Refactored code to use GameManager and PinCounter.
GameManager will be in charge of getting the action from ActionMaster by passing it a list of pinFalls. It will then call pinSetter's performAction with the action. Last, it will reset the ball.

PinCounter is replacing the script that was initially on LaneBox. It will keep track of the number of standing pins once the ball has exited the lane box and call PinsHaveSettled() once the number of standing pins remains the same after 3 seconds.